user data access moved to mongo assembly
